refactor: isolate some df code
